+++ src/api/routes.ts
@@ -285,6 +285,12 @@ export const setupRoutes = (server: Server) => {
 
   server.route({
     method: "GET",
+    path: "/collections/sources/v1",
+    options: collectionsEndpoints.getSourcesListingsV1Options,
+  });
+
+  server.route({
+    method: "GET",
     path: "/collections/v1",
     options: collectionsEndpoints.getCollectionsV1Options,
   });
@@ -560,6 +566,12 @@ export const setupRoutes = (server: Server) => {
     method: "GET",
     path: "/oracle/collections/{collection}/floor-ask/v2",
     options: oracleEndpoints.getCollectionFloorAskOracleV2Options,
+  });
+
+  server.route({
+    method: "GET",
+    path: "/oracle/collections/{collection}/floor-ask/v3",
+    options: oracleEndpoints.getCollectionFloorAskOracleV3Options,
   });
 
   // Orders