+++ src/models/api-keys/index.ts
@@ -14,6 +14,7 @@ export type ApiKeyRecord = {
   app_name: string;
   website: string;
   email: string;
+  tier: number;
   key?: string;
 };
 
@@ -38,7 +39,7 @@ export class ApiKeyManager {
     // Create the record in the database
     try {
       await idb.none(
-        "INSERT INTO api_keys (${this:name}) values (${this:csv}) ON CONFLICT DO NOTHING",
+        "INSERT INTO api_keys (${this:name}) VALUES (${this:csv}) ON CONFLICT DO NOTHING",
         values
       );
     } catch (e) {
@@ -93,7 +94,9 @@ export class ApiKeyManager {
         if (apiKey == "empty") {
           return null;
         } else {
-          return new ApiKeyEntity(JSON.parse(apiKey));
+          const apiKeyEntity = new ApiKeyEntity(JSON.parse(apiKey));
+          ApiKeyManager.apiKeys.set(key, apiKeyEntity); // Set in local memory storage
+          return apiKeyEntity;
         }
       } else {
         // check if it exists in the database
@@ -101,9 +104,9 @@ export class ApiKeyManager {
 
         if (fromDb) {
           Promise.race([redis.set(redisKey, JSON.stringify(fromDb)), timeout]); // Set in redis (no need to wait)
-          const apiKey = new ApiKeyEntity(fromDb);
-          ApiKeyManager.apiKeys.set(key, apiKey); // Set in local memory storage
-          return apiKey;
+          const apiKeyEntity = new ApiKeyEntity(fromDb);
+          ApiKeyManager.apiKeys.set(key, apiKeyEntity); // Set in local memory storage
+          return apiKeyEntity;
         } else {
           const pipeline = redis.pipeline();
           pipeline.set(redisKey, "empty");